refactor: Extract LanguageFunctionsSettings
